+++ modules/node-groups/main.tf
@@ -16,7 +16,6 @@ module "system_node_groups" {
   cluster_id                   = var.cluster_id
   cluster_version              = var.cluster_version
   cluster_version_full         = var.cluster_version_full
-  cluster_patch_upgrade        = var.cluster_patch_upgrade
   cni                          = var.cni
   fips                         = var.fips
   subnet_id                    = var.subnet_id
@@ -68,7 +67,6 @@ module "user_node_groups" {
   cluster_id                   = var.cluster_id
   cluster_version              = var.cluster_version
   cluster_version_full         = var.cluster_version_full
-  cluster_patch_upgrade        = var.cluster_patch_upgrade
   cni                          = var.cni
   fips                         = var.fips
   subnet_id                    = var.subnet_id