+++ modules/node-groups/modules/bootstrap-node-group-hack/scripts/delete-default-node-pool.sh
@@ -3,12 +3,12 @@ set -euo pipefail
 
 if [[ -n "${AZURE_ENVIRONMENT:-}" ]]
 then
-  az cloud set --name "${AZURE_ENVIRONMENT}"
+  az cloud set --name "${AZURE_ENVIRONMENT}" > /dev/null 2>&1
 fi
 
 if [[ -n "${AZURE_TENANT_ID:-}" ]] && [[ -n "${AZURE_CLIENT_ID:-}" ]] && [[ -n "${AZURE_CLIENT_SECRET:-}" ]]
 then
-  az login --service-principal --user "${AZURE_CLIENT_ID}"  --password "${AZURE_CLIENT_SECRET}" --tenant "${AZURE_TENANT_ID}"
+  az login --service-principal --user "${AZURE_CLIENT_ID}" --password "${AZURE_CLIENT_SECRET}" --tenant "${AZURE_TENANT_ID}" > /dev/null 2>&1
 fi
 
 lookup="$(az aks nodepool list \
@@ -28,5 +28,5 @@ fi
 
 if [[ -n "${AZURE_TENANT_ID:-}" ]] && [[ -n "${AZURE_CLIENT_ID:-}" ]] && [[ -n "${AZURE_CLIENT_SECRET:-}" ]]
 then
-  az logout --username "${AZURE_CLIENT_ID}"
+  az logout --username "${AZURE_CLIENT_ID}" > /dev/null 2>&1 || true
 fi